requirements.txt
Django==2.1.3
mysqlclient==1.3.10
python-frontmatter==0.4.2
